refactor: move Parser to terraform package
