+++ internal/terraform/parser_test.go
@@ -1,4 +1,4 @@
-package tfmod
+package terraform
 
 import (
 	"os"
@@ -7,7 +7,6 @@ import (
 	"testing"
 
 	"github.com/tmknom/tfmod/internal/dir"
-	"github.com/tmknom/tfmod/internal/terraform"
 )
 
 func TestParser_Parse(t *testing.T) {
@@ -59,7 +58,7 @@ func (s *FakeParserStore) Actual() []string {
 	return s.list
 }
 
-func (s *FakeParserStore) Save(moduleDir *terraform.ModuleDir, tfDir *terraform.TfDir) {
+func (s *FakeParserStore) Save(moduleDir *ModuleDir, tfDir *TfDir) {
 	pair := strings.Join([]string{moduleDir.Rel(), tfDir.Rel()}, ":")
 	s.list = append(s.list, pair)
 }
